Move fileSystem package into tsfile package to be consistent with other packages
